+++ src/server/index.js
@@ -1,3 +1,4 @@
 export { GameServer, createGameServer, createServer } from "./server.js";
 export { World, WorldError, xpRequiredForLevel } from "./world.js";
+export { PROTOCOL, validate as validateProtocolMessage } from "./protocol.js";
 export * from "./definitions.js";